Sorted the images and annotations so they line up; Developed and initialized the model (from PyTorch's segmentation models-temporarily); set loss fn and optimizer; rough drafted the train and eval function;
When parsing the annotations file, I am not sure how to organize it so that it stores the data in a list where the first index of annotations corresponds to that image. I want to do this so I can split up the data in to train, test, and dataloaders at a later stage.